First pass at an object storage (swift) client.
It's called noox for fun. Its mocked counterpart is called moox.

This is still very rough, the moox response codes and error messages will
need to be adapted to resemble the real world codes/messages more closely.

Auhtentication on OpenStack Object Storage is handled quite differently
than on S3. On S3 you basically sign every request with your credentials,
so there is no separate authentication request.
On Object Storage, you first call an authentication endpoint, which responds
with a security token that is only valid for 24 hours and which you are supposed
to included in headers in every subsequent request, as well as the base URL
pointing to your storage service.

Since creating a token is an asynchronous operation which needs to be done
before any other requests; and because tokens expire, there are some differences
in how the noox client is used compared to nox.

Noox tries to be low level, so you are responsible for authenticating with
the endpoint, as well as for keeping track of token expiry, calling the
authenticate function at least every 24 hours.
I am not completely sold on this idea yet, so noox might do automatic
authentication in the future, but this was easier to implement as the
first step.