Use new Node 0.10 stream base classes in mo(o)x.
Node 0.10 provides a set of base classes for implementing streams.
Use that instead of building fake readable/writable streams from scratch.
This makes the fake requests/responses in mox/moox comply to the new streams API.

This only works with Node 0.10+. It could be made backwards compatible,
but I don't have any need for that at the moment.